Extract sal method into separate module
