+++ pysteps/verification/interface.py
@@ -166,7 +166,8 @@ def get_method(name, type="deterministic"):
 
         from .detcatscores import det_cat_fct
         from .detcontscores import det_cont_fct
-        from .spatialscores import binary_mse, fss, sal
+        from .spatialscores import binary_mse, fss
+        from .sal import sal
 
         # categorical
         if name in [
